Movement improvements

+++ Assets/Artemis.Sample/Settings/Configuration.cs
@@ -7,6 +7,6 @@ namespace Artemis.Sample
         public const short ServerPort = 12345;
         public const int TicksPerSecond = 8;
         public static readonly TimeSpan TickInterval = TimeSpan.FromSeconds(1f / TicksPerSecond);
-        public const float PlayerMovementSpeed = 2;
+        public const double PlayerMovementSpeed = 2;
     }
 }
